feat: rewrite in type checked JS
