+++ .eslintrc.cjs
@@ -1,13 +1,16 @@
-// @ts-check
-
 'use strict';
 
-/** @type import('eslint').ESLint.ConfigData */
+/** @type {import('eslint').ESLint.ConfigData} */
 module.exports = {
   root: true,
   extends: ['@hatena/hatena-stable', '@hatena/hatena-stable/+typescript', '@hatena/hatena-stable/+prettier'],
   env: {
     node: true,
+  },
+  settings: {
+    'import/resolver': {
+      typescript: {},
+    },
   },
   rules: {
     'import/no-default-export': 0,
@@ -19,14 +22,6 @@ module.exports = {
       files: ['*.cjs'],
       parserOptions: {
         sourceType: 'script',
-      },
-    },
-    {
-      files: ['*.ts'],
-      settings: {
-        'import/resolver': {
-          typescript: {},
-        },
       },
     },
   ],
