+++ .gitignore
@@ -1,5 +1,3 @@
-/lib
-
 ### https://raw.github.com/github/gitignore/4488915eec0b3a45b5c63ead28f286819c0917de/Node.gitignore
 
 # Logs
