+++ lib/classic/index.js
@@ -1,0 +1,13 @@
+import javascript from './javascript.js';
+import prettier from './prettier.js';
+import react from './react.js';
+import typescript from './typescript.js';
+
+const config = {
+  javascript,
+  prettier,
+  react,
+  typescript,
+};
+
+export default config;
