+++ lib/classic/prettier.js
@@ -1,0 +1,6 @@
+/** @type {import('eslint').ESLint.ConfigData} */
+const config = {
+  extends: ['prettier'],
+};
+
+export default config;
