+++ lib/classic/typescript.js
@@ -1,7 +1,7 @@
-import type { ESLint } from 'eslint';
-import rules from '../rules';
+import rules from '../rules/index.js';
 
-const config: ESLint.ConfigData = {
+/** @type {import('eslint').ESLint.ConfigData} */
+const config = {
   overrides: [
     {
       // *.js などではこれらのルールが適用されないようにする
