+++ lib/config/index.js
@@ -1,40 +1,32 @@
 import jsPlugin from '@eslint/js';
-import type { ESLint, Linter } from 'eslint';
 import prettierConfig from 'eslint-config-prettier';
 import importPlugin from 'eslint-plugin-import';
 import reactPlugin from 'eslint-plugin-react';
 import reactHooksPlugin from 'eslint-plugin-react-hooks';
 import globals from 'globals';
 import tsEslint from 'typescript-eslint';
-import rules from '../rules';
+import rules from '../rules/index.js';
 
-export type ConfigOptions = Readonly<
-  Partial<{
-    /**
-     * TypeScript の設定ファイル (languageOptions.parserOptions.project)
-     * デフォルト: ./tsconfig.json
-     */
-    tsProject: boolean | string | string[] | undefined;
-    /**
-     * React を使用するか. true の場合, React に関連する設定を有効にする
-     * デフォルト: false
-     */
-    react: boolean | undefined;
-    /**
-     * Prettier を使用するか. true の場合, eslint-config-prettier を使ってフォーマットに関するルールを無効化する
-     * デフォルト: true
-     */
-    prettier: boolean | undefined;
-  }>
->;
+/**
+ * @typedef ConfigOptions
+ * @property {boolean | string | string[] | undefined} [tsProject]
+ * TypeScript の設定ファイル (languageOptions.parserOptions.project)
+ * デフォルト: ./tsconfig.json
+ * @property {boolean | undefined} [react]
+ * React を使用するか. true の場合, React に関連する設定を有効にする
+ * デフォルト: false
+ * @property {boolean | undefined} [prettier]
+ * Prettier を使用するか. true の場合, eslint-config-prettier を使ってフォーマットに関するルールを無効化する
+ * デフォルト: true
+ */
 
 /**
  * ESLint の設定を作る
- * @param options オプション
- * @param configs カスタム設定の配列
- * @returns 設定の配列
+ * @param {ConfigOptions} [options] オプション
+ * @param {readonly import('eslint').Linter.FlatConfig[]} [configs] カスタム設定の配列
+ * @returns {import('eslint').Linter.FlatConfig[]} 設定の配列
  */
-export function config(options?: ConfigOptions, configs?: readonly Linter.FlatConfig[]): Linter.FlatConfig[] {
+export function config(options, configs) {
   const tsProject = options?.tsProject ?? './tsconfig.json';
   const react = options?.react ?? false;
   const prettier = options?.prettier ?? true;
@@ -43,7 +35,7 @@ export function config(options?: ConfigOptions, configs?: readonly Linter.FlatCo
     // # Linter 自体の設定
     {
       plugins: {
-        '@typescript-eslint': tsEslint.plugin as ESLint.Plugin,
+        '@typescript-eslint': /** @type {import('eslint').ESLint.Plugin} */ (tsEslint.plugin),
         'import': importPlugin,
         'react': reactPlugin,
         'react-hooks': reactHooksPlugin,
@@ -89,7 +81,7 @@ export function config(options?: ConfigOptions, configs?: readonly Linter.FlatCo
       files: ['**/*.{ts,tsx,cts,mts}'],
       languageOptions: {
         sourceType: 'module',
-        parser: tsEslint.parser as Linter.FlatConfigParserModule,
+        parser: /** @type {import('eslint').Linter.FlatConfigParserModule} */ (tsEslint.parser),
         parserOptions: tsProject ? { project: tsProject } : {},
       },
       settings: {
@@ -112,7 +104,7 @@ export function config(options?: ConfigOptions, configs?: readonly Linter.FlatCo
     ...map({ files: ['**/*.{ts,tsx,cts,mts}'] }, [
       // languageOptions なども含まれるが上で設定しているものと重複するので, ここでは rules のみに絞る
       ...tsEslint.configs.recommendedTypeChecked.flatMap((config) =>
-        config.rules ? [{ rules: config.rules as Linter.RulesRecord }] : [],
+        config.rules ? [{ rules: /** @type {import('eslint').Linter.RulesRecord} */ (config.rules) }] : [],
       ),
       { rules: importPlugin.configs.typescript.rules },
       { rules: rules.typescript },
@@ -125,6 +117,11 @@ export function config(options?: ConfigOptions, configs?: readonly Linter.FlatCo
   ];
 }
 
-function map(base: Linter.FlatConfig, configs: readonly Linter.FlatConfig[]): Linter.FlatConfig[] {
+/**
+ * @param {import('eslint').Linter.FlatConfig} base
+ * @param {readonly import('eslint').Linter.FlatConfig[]} configs
+ * @returns {import('eslint').Linter.FlatConfig[]}
+ */
+function map(base, configs) {
   return configs.map((config) => ({ ...base, ...config }));
 }
