+++ lib/index.js
@@ -1,0 +1,5 @@
+import classicConfig from './classic/index.js';
+import { config } from './config/index.js';
+
+export { config, classicConfig as __internal_classicConfig };
+export default config;
