+++ lib/rules/index.js
@@ -1,0 +1,12 @@
+import javascript from './javascript.js';
+import react, { typescript as typescriptReact } from './react.js';
+import typescript from './typescript.js';
+
+const rules = {
+  javascript,
+  react,
+  typescript,
+  typescriptReact,
+};
+
+export default rules;
