+++ lib/rules/javascript.js
@@ -1,6 +1,5 @@
-import type { Linter } from 'eslint';
-
-const rules: Linter.RulesRecord = {
+/** @type {import('eslint').Linter.RulesRecord} */
+const rules = {
   // # eslint
   'no-var': 2,
   // コーディングスタイル統一のため、`const fn = function() { ... }` 形式の関数定義を禁止する。
