+++ lib/rules/typescript.js
@@ -1,6 +1,5 @@
-import type { Linter } from 'eslint';
-
-const rules: Linter.RulesRecord = {
+/** @type {import('eslint').Linter.RulesRecord} */
+const rules = {
   // # eslint
   'strict': 0,
