--- src/classic/index.ts
@@ -1,13 +1,0 @@
-import javascript from './javascript';
-import prettier from './prettier';
-import react from './react';
-import typescript from './typescript';
-
-const config = {
-  javascript,
-  prettier,
-  react,
-  typescript,
-};
-
-export default config;
