--- src/classic/prettier.ts
@@ -1,7 +1,0 @@
-import type { ESLint } from 'eslint';
-
-const config: ESLint.ConfigData = {
-  extends: ['prettier'],
-};
-
-export default config;
