--- src/index.ts
@@ -1,6 +1,0 @@
-import classicConfig from './classic';
-import type { ConfigOptions } from './config';
-import { config } from './config';
-
-export { ConfigOptions, config, classicConfig as __internal_classicConfig };
-export default config;
