--- src/rules/index.ts
@@ -1,12 +1,0 @@
-import javascript from './javascript';
-import react, { typescript as typescriptReact } from './react';
-import typescript from './typescript';
-
-const rules = {
-  javascript,
-  react,
-  typescript,
-  typescriptReact,
-};
-
-export default rules;
